+++ packages/jaspr_content/lib/src/data_loader/filesystem_data_loader.dart
@@ -23,7 +23,8 @@ class FilesystemDataLoader implements DataLoader {
     );
   }
 
-  static DirectoryWatcher _defaultWatcherFactory(String path) => DirectoryWatcher(path);
+  static DirectoryWatcher _defaultWatcherFactory(String path) =>
+      DirectoryWatcher(path);
 
   static final Map<String, FilesystemDataLoader> _instance = {};
 
@@ -91,8 +92,10 @@ class FilesystemDataLoader implements DataLoader {
       await for (final entity in dir.list()) {
         if (entity is File) {
           final name = entity.path.split(fileSystem.path.separator).last;
-          final key = name.split('.').first;
-          data[key] = DataLoader.parseData(name, await entity.readAsString());
+          if (DataLoader.canParse(name)) {
+            final key = name.split('.').first;
+            data[key] = DataLoader.parseData(name, await entity.readAsString());
+          }
         } else if (entity is Directory) {
           final key = entity.path.split(fileSystem.path.separator).last;
           data[key] = await _loadData(entity);